% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.coverage}
\alias{.coverage}
\title{Compute simple coverage summary}
\usage{
.coverage(items_with_status)
}
\arguments{
\item{items_with_status}{A data frame containing a \code{status} column with
values in \code{c("reported", "missing", "na")}.}
}
\value{
A list with counts and a percentage of applicable items reported.
}
\description{
Compute simple coverage summary
}
\keyword{internal}
