% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_report.R
\name{plot.invert_report}
\alias{plot.invert_report}
\title{Plot method for invert_report}
\usage{
\method{plot}{invert_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{invert_report}.}

\item{...}{Unused.}
}
\description{
Plot method for invert_report
}
