% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_paper}
\alias{validate_paper}
\title{Validate paper metadata}
\usage{
validate_paper(paper)
}
\arguments{
\item{paper}{A named list.}
}
\value{
Invisibly, \code{TRUE} if valid. Errors otherwise.
}
\description{
Validate paper metadata
}
\keyword{internal}
